refactor(FeedbackRuleEvaluator): Evaluate multiple students in new class
#1252
