+++ src/assets/wise5/components/common/feedbackRule/FeedbackRuleEvaluator.ts
@@ -5,12 +5,13 @@ import { FeedbackRuleExpression } from './FeedbackRuleExpression';
 import { TermEvaluator } from './TermEvaluator/TermEvaluator';
 import { TermEvaluatorFactory } from './TermEvaluator/TermEvaluatorFactory';
 
-export class FeedbackRuleEvaluator {
+export class FeedbackRuleEvaluator<T extends CRaterResponse | CRaterResponse[]> {
   defaultFeedback = $localize`Thanks for submitting your response.`;
+  protected factory = new TermEvaluatorFactory();
 
-  constructor(private component: FeedbackRuleComponent) {}
+  constructor(protected component: FeedbackRuleComponent) {}
 
-  getFeedbackRule(response: CRaterResponse | CRaterResponse[]): FeedbackRule {
+  getFeedbackRule(response: T): FeedbackRule {
     for (const feedbackRule of this.component.getFeedbackRules()) {
       if (this.satisfiesRule(response, Object.assign(new FeedbackRule(), feedbackRule))) {
         return feedbackRule;
@@ -19,28 +20,13 @@ export class FeedbackRuleEvaluator {
     return this.getDefaultRule(this.component.getFeedbackRules());
   }
 
-  getFeedbackRules(response: CRaterResponse[]): FeedbackRule[] {
-    const matchedRules = this.component
-      .getFeedbackRules()
-      .filter((rule) => this.satisfiesRule(response, Object.assign(new FeedbackRule(), rule)));
-    return matchedRules.length > 0
-      ? matchedRules
-      : [this.getDefaultRule(this.component.getFeedbackRules())];
-  }
-
-  private satisfiesRule(
-    response: CRaterResponse | CRaterResponse[],
-    feedbackRule: FeedbackRule
-  ): boolean {
+  protected satisfiesRule(response: T, feedbackRule: FeedbackRule): boolean {
     return this.isSpecialRule(feedbackRule)
       ? this.satisfiesSpecialRule(response, feedbackRule)
       : this.satisfiesSpecificRule(response, feedbackRule);
   }
 
-  private satisfiesSpecialRule(
-    response: CRaterResponse | CRaterResponse[],
-    feedbackRule: FeedbackRule
-  ): boolean {
+  private satisfiesSpecialRule(response: T, feedbackRule: FeedbackRule): boolean {
     return (
       this.satisfiesNonScorableRule(response, feedbackRule) ||
       this.satisfiesFinalSubmitRule(response, feedbackRule) ||
@@ -48,59 +34,37 @@ export class FeedbackRuleEvaluator {
     );
   }
 
-  private satisfiesFinalSubmitRule(
-    response: CRaterResponse | CRaterResponse[],
-    feedbackRule: FeedbackRule
-  ): boolean {
+  protected satisfiesFinalSubmitRule(response: T, feedbackRule: FeedbackRule): boolean {
     return (
       this.hasMaxSubmitAndIsFinalSubmitRule(feedbackRule) &&
-      (response instanceof CRaterResponse
-        ? this.component.hasMaxSubmitCountAndUsedAllSubmits(response.submitCounter)
-        : response.some((response: CRaterResponse) => {
-            return this.component.hasMaxSubmitCountAndUsedAllSubmits(response.submitCounter);
-          }))
+      this.component.hasMaxSubmitCountAndUsedAllSubmits((response as CRaterResponse).submitCounter)
     );
   }
 
-  private hasMaxSubmitAndIsFinalSubmitRule(feedbackRule: FeedbackRule): boolean {
+  protected hasMaxSubmitAndIsFinalSubmitRule(feedbackRule: FeedbackRule): boolean {
     return this.component.hasMaxSubmitCount() && FeedbackRule.isFinalSubmitRule(feedbackRule);
   }
 
-  private satisfiesSecondToLastSubmitRule(
-    response: CRaterResponse | CRaterResponse[],
-    feedbackRule: FeedbackRule
-  ): boolean {
+  protected satisfiesSecondToLastSubmitRule(response: T, feedbackRule: FeedbackRule): boolean {
     return (
       this.hasMaxSubmitAndIsSecondToLastSubmitRule(feedbackRule) &&
-      (response instanceof CRaterResponse
-        ? this.isSecondToLastSubmit(response.submitCounter)
-        : response.some((response: CRaterResponse) => {
-            return this.isSecondToLastSubmit(response.submitCounter);
-          }))
+      this.isSecondToLastSubmit((response as CRaterResponse).submitCounter)
     );
   }
 
-  private hasMaxSubmitAndIsSecondToLastSubmitRule(feedbackRule: FeedbackRule): boolean {
+  protected hasMaxSubmitAndIsSecondToLastSubmitRule(feedbackRule: FeedbackRule): boolean {
     return (
       this.component.hasMaxSubmitCount() && FeedbackRule.isSecondToLastSubmitRule(feedbackRule)
     );
   }
 
-  private isSecondToLastSubmit(submitCounter: number): boolean {
+  protected isSecondToLastSubmit(submitCounter: number): boolean {
     return this.component.getNumberOfSubmitsLeft(submitCounter) === 1;
   }
 
-  private satisfiesNonScorableRule(
-    response: CRaterResponse | CRaterResponse[],
-    feedbackRule: FeedbackRule
-  ): boolean {
+  protected satisfiesNonScorableRule(response: T, feedbackRule: FeedbackRule): boolean {
     return (
-      feedbackRule.expression === 'isNonScorable' &&
-      (response instanceof CRaterResponse
-        ? response.isNonScorable()
-        : response.some((response: CRaterResponse) => {
-            return response.isNonScorable();
-          }))
+      feedbackRule.expression === 'isNonScorable' && (response as CRaterResponse).isNonScorable()
     );
   }
 
@@ -110,10 +74,7 @@ export class FeedbackRuleEvaluator {
     );
   }
 
-  private satisfiesSpecificRule(
-    response: CRaterResponse | CRaterResponse[],
-    feedbackRule: FeedbackRule
-  ): boolean {
+  private satisfiesSpecificRule(response: T, feedbackRule: FeedbackRule): boolean {
     const termStack = [];
     for (const term of feedbackRule.getPostfixExpression()) {
       if (FeedbackRuleExpression.isOperand(term)) {
@@ -128,11 +89,7 @@ export class FeedbackRuleEvaluator {
     return true;
   }
 
-  private evaluateOperator(
-    operator: string,
-    termStack: string[],
-    response: CRaterResponse | CRaterResponse[]
-  ) {
+  private evaluateOperator(operator: string, termStack: string[], response: T) {
     if (this.evaluateOperatorExpression(operator, termStack, response)) {
       termStack.push('true');
     } else {
@@ -140,11 +97,7 @@ export class FeedbackRuleEvaluator {
     }
   }
 
-  private evaluateOperatorExpression(
-    operator: string,
-    termStack: string[],
-    response: CRaterResponse | CRaterResponse[]
-  ): boolean {
+  private evaluateOperatorExpression(operator: string, termStack: string[], response: T): boolean {
     if (['&&', '||'].includes(operator)) {
       return this.evaluateAndOrExpression(operator, termStack, response);
     } else {
@@ -152,11 +105,7 @@ export class FeedbackRuleEvaluator {
     }
   }
 
-  private evaluateAndOrExpression(
-    operator: string,
-    termStack: string[],
-    response: CRaterResponse | CRaterResponse[]
-  ): boolean {
+  private evaluateAndOrExpression(operator: string, termStack: string[], response: T): boolean {
     const term1 = termStack.pop();
     const term2 = termStack.pop();
     return operator === '&&'
@@ -164,24 +113,16 @@ export class FeedbackRuleEvaluator {
       : this.evaluateTerm(term1, response) || this.evaluateTerm(term2, response);
   }
 
-  private evaluateNotExpression(
-    termStack: string[],
-    response: CRaterResponse | CRaterResponse[]
-  ): boolean {
+  private evaluateNotExpression(termStack: string[], response: T): boolean {
     return !this.evaluateTerm(termStack.pop(), response);
   }
 
-  private evaluateTerm(term: string, response: CRaterResponse | CRaterResponse[]): boolean {
-    const factory = new TermEvaluatorFactory();
-    const evaluator: TermEvaluator = factory.getTermEvaluator(term);
-    return response instanceof CRaterResponse
-      ? evaluator.evaluate(response)
-      : response.some((response: CRaterResponse) => {
-          return evaluator.evaluate(response);
-        });
+  protected evaluateTerm(term: string, response: T): boolean {
+    const evaluator: TermEvaluator = this.factory.getTermEvaluator(term);
+    return evaluator.evaluate(response as CRaterResponse);
   }
 
-  private getDefaultRule(feedbackRules: FeedbackRule[]): FeedbackRule {
+  protected getDefaultRule(feedbackRules: FeedbackRule[]): FeedbackRule {
     return (
       feedbackRules.find((rule) => FeedbackRule.isDefaultRule(rule)) ||
       Object.assign(new FeedbackRule(), {
